+++ src/OnBoard/OnBoardSteps/StepManualZones1.php
@@ -14,6 +14,7 @@ namespace Invertus\dpdBaltics\OnBoard\OnBoardSteps;
 
 use DPDBaltics;
 use Invertus\dpdBaltics\Config\Config;
+use Invertus\dpdBaltics\Infrastructure\Bootstrap\ModuleTabs;
 use Invertus\dpdBaltics\OnBoard\AbstractOnBoardStep;
 use Invertus\dpdBaltics\OnBoard\Objects\OnBoardButton;
 use Invertus\dpdBaltics\OnBoard\Objects\OnBoardFastMoveButton;
@@ -67,7 +68,7 @@ class StepManualZones1 extends AbstractOnBoardStep
     public function takeStepAction()
     {
         if ($this->stepActionService->nextStepIfRightController(
-            DPDBaltics::ADMIN_ZONES_CONTROLLER,
+            ModuleTabs::ADMIN_ZONES_CONTROLLER,
             Config::STEP_MANUAL_ZONES_2
         )) {
             $this->stepActionService->setManualConfigCompletedSteps(Config::ON_BOARD_ZONES_PART);